backend/requirements.txt
# File: /backend/requirements.txt

# Core Framework
fastapi==0.68.1
uvicorn[standard]==0.15.0
gunicorn==20.1.0
python-multipart==0.0.5

# Scraping
scrapy==2.5.1
selenium==4.1.0
undetected-chromedriver==3.1.5
beautifulsoup4==4.9.3
aiohttp==3.8.1
price-parser==0.3.4

# Database
sqlalchemy==1.4.23
alembic==1.7.3
psycopg2-binary==2.9.1
aiosqlite==0.17.0

# Image Processing
Pillow==8.3.2
python-magic==0.4.24
aiofiles==0.7.0

# Utils
pydantic==1.8.2
python-dotenv==0.19.0
tenacity==8.0.1
aiocache==0.11.1

# Monitoring
prometheus-client==0.11.0
psutil==5.8.0

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1

# Development
black==21.7b0
flake8==3.9.2
mypy==0.910
isort==5.9.3